refactor: rename challenge to challenge1
