+++ src/arrayFunctions/challenge1.ts
